requirements.txt
asgiref==3.10.0
asttokens==3.0.0
bcrypt==5.0.0
colorama==0.4.6
decorator==5.2.1
Django==5.2.7
django-extensions==4.1
djangorestframework==3.16.1
executing==2.2.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
isort==7.0.0
jedi==0.19.2
matplotlib-inline==0.2.1
parso==0.8.5
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.10.1
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
tzdata==2025.2
wcwidth==0.2.14
python-dotenv